update README, gitignore, config files, code cleaning

+++ functions.py
@@ -19,13 +19,14 @@ MODE = 'web'
 def clean_images(dirpath):
     print('cleaning images...')
     for the_file in os.listdir(dirpath):
-        file_path = os.path.join(dirpath, the_file)
-        try:
-            if os.path.isfile(file_path):
-                os.unlink(file_path)
-            #elif os.path.isdir(file_path): shutil.rmtree(file_path)
-        except Exception as e:
-            print(e)
+        if the_file != 'index.php':
+            file_path = os.path.join(dirpath, the_file)
+            try:
+                if os.path.isfile(file_path):
+                    os.unlink(file_path)
+                #elif os.path.isdir(file_path): shutil.rmtree(file_path)
+            except Exception as e:
+                print(e)
 
 # PIL stuff
 def compare(file1, file2):
